add deduped institutions to JATS
